ikev2: Don't treat delayed IKE_SA_INIT retransmits as fatal error
If the IKE_AUTH message has already been handled we expect the next
message ID to be 2.  So if a retransmit of the IKE_SA_INIT with MID 0
arrives delayed we don't detect it as retransmit (we only recognize the
one we most recently responded to, i.e. expected MID - 1).  Therefore,
instead of just ignoring the retransmitted IKE_SA_INIT the existing
check caused the IKE_SA to get destroyed.  This change ensures that
the check actually does what the comment indicates, i.e. make sure
the MID of the IKE_SA_INIT is valid.

 #816