tls-peer: Use KEY_ID to find trusted public key for peer
In case a CA certificate uses the same subject DN as the server the
previous code could end up trying to verify the server's signature with
the CA certificate's public key.  By using the KEY_ID instead of the
subject DN we make sure to find the right certificate (unless the same
key is used for the CA and the server).

 #849